# backend/src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Football Website API",
    "version": "1.0.0",
    "description": "Public read API (v1) for news, matches, teams, players, competitions and transfers. All list responses are paginated. Errors use a stable { error: { code, message } } envelope."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [],
  "paths": {
...
    "/matches": {
      "get": {
        "summary": "List matches",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 20
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "scheduled",
                "pre_match",
                "live",
                "half_time",
                "extra_time",
                "penalty_shootout",
                "finished",
                "postponed",
                "cancelled",
                "abandoned",
                "suspended"
              ]
            }
          },
          {
            "name": "phase",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "upcoming",
                "live",
                "finished"
              ]
            }
          },
          {
            "name": "competition",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "season",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated matches"
          },
          "400": {
            "description": "Invalid parameters"
          }
        }
      }
    },
    "/matches/{slug}": {
      "get": {
        "summary": "Get match by slug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Match"
          },
          "400": {
            "description": "Invalid slug"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/media/{id}": {
      "get": {
        "summary": "Media metadata + variants (public only when referenced by published articles)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Media with variants"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Private media"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "summary": "Update media metadata alt/caption/credit (owner or editor)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated media"
          }
        }
      },
      "delete": {
        "summary": "Safe delete (409 while referenced by articles; otherwise storage+rows removed)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "409": {
            "description": "Referenced media"
          }
        }
      }
    },
...
    "/players": {
      "get": {
        "summary": "List players",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 20
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "position",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nationality",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated players"
          },
          "400": {
            "description": "Invalid parameters"
          }
        }
      }
    },
    "/players/{slug}": {
      "get": {
        "summary": "Get player by slug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Player"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}